+++ test/xchain_crypto/utils_test.dart
@@ -4,7 +4,7 @@ import 'package:xchain_dart/src/xchain_crypto/utils.dart';
 void main() {
   test('is empty', () {
     String source = '';
-    List<Address> addresses = [];
+    List<AssetAddress> addresses = [];
     try {
       addresses = substractAddress(source);
     } catch (err) {
@@ -26,7 +26,7 @@ void main() {
 
   test('starts with bitcoin chain prefix', () {
     String source = 'bitcoin:3QaesQ25kJc4tyCQM5wJ54ky39DNsUMx7Z';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(addresses.length, 1);
     expect(addresses.first.address, '3QaesQ25kJc4tyCQM5wJ54ky39DNsUMx7Z');
     expect(addresses.first.asset, 'BTC.BTC');
@@ -35,7 +35,7 @@ void main() {
 
   test('starts with bitcoin chain prefix', () {
     String source = 'BITCOIN:BC1QFPFPW7DLQT3AZCYQXMCUZM5L2XRM6PD6ZVSE9Q';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(addresses.length, 1);
     expect(
         addresses.first.address, 'BC1QFPFPW7DLQT3AZCYQXMCUZM5L2XRM6PD6ZVSE9Q');
@@ -45,7 +45,7 @@ void main() {
 
   test('starts with bitcoin cash chain prefix', () {
     String source = 'bitcoincash:qpl4lfjq7emfg8p4akr6p27dap5duj35zcc82aqul5';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(addresses.length, 1);
     expect(
         addresses.first.address, 'qpl4lfjq7emfg8p4akr6p27dap5duj35zcc82aqul5');
@@ -55,7 +55,7 @@ void main() {
 
   test('starts with ethereum chain prefix', () {
     String source = 'ethereum:0xC52A857FDa38994CB6CC8e0DE2AEDD67a7353e0d';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(
         addresses.first.address, '0xC52A857FDa38994CB6CC8e0DE2AEDD67a7353e0d');
     expect(addresses.first.asset, 'ETH.ETH');
@@ -64,7 +64,7 @@ void main() {
 
   test('starts with dragonsdex chain prefix', () {
     String source = 'dragonsdex:0xC52A857FDa38994CB6CC8e0DE2AEDD67a7353e0d';
-    List<Address> addresses = [];
+    List<AssetAddress> addresses = [];
     try {
       addresses = substractAddress(source);
     } catch (err) {
@@ -76,7 +76,7 @@ void main() {
 
   test('starts with binance chain prefix', () {
     String source = 'binance:bnb1vxyxxkqdke8r55r6fzhprtj8qwgecudj0h5svr';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(
         addresses.first.address, 'bnb1vxyxxkqdke8r55r6fzhprtj8qwgecudj0h5svr');
     expect(addresses.first.asset, 'BNB.BNB');
@@ -85,7 +85,7 @@ void main() {
 
   test('bitcoin legacy address without chain prefix', () {
     String source = '16we9adsewmBDKv5CSgeRMZPo3RadcgVZV';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(addresses.length, 2);
     expect(addresses.first.address, '16we9adsewmBDKv5CSgeRMZPo3RadcgVZV');
     expect(addresses[0].asset, 'BCH.BCH');
@@ -95,7 +95,7 @@ void main() {
 
   test('bitcoin segwit address without chain prefix', () {
     String source = '3QaesQ25kJc4tyCQM5wJ54ky39DNsUMx7Z';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(addresses.length, 2);
     expect(addresses.first.address, '3QaesQ25kJc4tyCQM5wJ54ky39DNsUMx7Z');
     expect(addresses[0].asset, 'BTC.BTC');
@@ -105,7 +105,7 @@ void main() {
 
   test('bitcoin native segwit address without chain prefix', () {
     String source = 'bc1qfw00pnu77vvw3r8fpterjukx0u3nj26n724pq3';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(addresses.length, 1);
     expect(
         addresses.first.address, 'bc1qfw00pnu77vvw3r8fpterjukx0u3nj26n724pq3');
@@ -115,7 +115,7 @@ void main() {
 
   test('ethereum address without chain prefix', () {
     String source = '0xC52A857FDa38994CB6CC8e0DE2AEDD67a7353e0d';
-    List<Address> addresses = substractAddress(source);
+    List<AssetAddress> addresses = substractAddress(source);
     expect(
         addresses.first.address, '0xC52A857FDa38994CB6CC8e0DE2AEDD67a7353e0d');
     expect(addresses.first.asset, 'ETH.ETH');
@@ -124,7 +124,7 @@ void main() {
 
   test('unsupported Chain address', () {
     String source = 'AJXPYa2aYizxXfhcEcmom1xuEyZLF6DX5b';
-    List<Address> addresses = [];
+    List<AssetAddress> addresses = [];
     try {
       addresses = substractAddress(source);
     } catch (err) {